backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Todo application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "alembic>=1.14.0",
    "pyjwt>=2.10.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.1.0",
    "openai>=1.59.3",
    "mcp>=1.1.2",
    "google-genai>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
